customer_support_agent/requirements.txt
openai
faiss-cpu
numpy
tiktoken
